Refactor tpu_feedback code to its own module
