+++ tpu-client-next/src/connection_worker.rs
@@ -5,15 +5,13 @@ use {
     super::SendTransactionStats,
     crate::{
         quic_networking::send_data_over_stream, send_transaction_stats::record_error,
-        transaction_batch::TransactionBatch,
+        tpu_feedback::recv_tpu_feedback, transaction_batch::TransactionBatch,
     },
     log::*,
     quinn::{ConnectError, Connection, Endpoint},
     solana_clock::{DEFAULT_MS_PER_SLOT, MAX_PROCESSING_AGE, NUM_CONSECUTIVE_LEADER_SLOTS},
     solana_measure::measure::Measure,
-    solana_streamer::tpu_feedback::{
-        TpuFeedback, TYPE_PRIORITY_FEES, TYPE_TIMESTAMP, TYPE_TRANSACTION_STATE, TYPE_VERSION,
-    },
+    solana_streamer::tpu_feedback::TpuFeedback,
     solana_time_utils::timestamp,
     std::{
         net::SocketAddr,
@@ -46,70 +44,6 @@ enum ConnectionState {
     Active(Connection),
     Retry(usize),
     Closing,
-}
-
-fn parse_feedback(data: &[u8]) -> Option<TpuFeedback> {
-    let mut offset = 0;
-    let mut version: Option<u8> = None;
-    let mut timestamp: Option<u64> = None;
-    let mut transactions: Vec<(u64, u32)> = vec![];
-    let mut priority_fees: Option<(u64, u64, u64)> = None;
-
-    while offset + 3 <= data.len() {
-        let t = data[offset];
-        let l = u16::from_le_bytes([data[offset + 1], data[offset + 2]]) as usize;
-        offset += 3;
-
-        if offset + l > data.len() {
-            break;
-        }
-
-        match t {
-            TYPE_VERSION if l == 1 => {
-                version = Some(data[offset]);
-            }
-            TYPE_TIMESTAMP if l == 8 => {
-                timestamp = Some(u64::from_le_bytes(
-                    data[offset..offset + 8].try_into().ok()?,
-                ));
-            }
-            TYPE_TRANSACTION_STATE => {
-                let mut sub_offset = 0;
-                while sub_offset + 12 <= l {
-                    let tx_sig = u64::from_le_bytes(
-                        data[offset + sub_offset..offset + sub_offset + 8]
-                            .try_into()
-                            .ok()?,
-                    );
-                    let state = u32::from_le_bytes(
-                        data[offset + sub_offset + 8..offset + sub_offset + 12]
-                            .try_into()
-                            .ok()?,
-                    );
-                    transactions.push((tx_sig, state));
-                    sub_offset += 12;
-                }
-            }
-            TYPE_PRIORITY_FEES if l == 24 => {
-                let min = u64::from_le_bytes(data[offset..offset + 8].try_into().ok()?);
-                let median = u64::from_le_bytes(data[offset + 8..offset + 16].try_into().ok()?);
-                let max = u64::from_le_bytes(data[offset + 16..offset + 24].try_into().ok()?);
-                priority_fees = Some((min, median, max));
-            }
-            _ => {
-                debug!("Unknown or malformed TLV type: {} with length {}", t, l);
-            }
-        }
-
-        offset += l;
-    }
-
-    Some(TpuFeedback {
-        version: version?,
-        timestamp: timestamp?,
-        transactions,
-        priority_fees: priority_fees?,
-    })
 }
 
 impl Drop for ConnectionState {
@@ -266,46 +200,6 @@ impl ConnectionWorker {
         );
     }
 
-    async fn recv_datagrams(
-        connection: Connection,
-        cancel: CancellationToken,
-        sender: broadcast::Sender<TpuFeedback>,
-    ) {
-        loop {
-            tokio::select! {
-                _ = cancel.cancelled() => {
-                    debug!("Recv datagrams task cancelled.");
-                    break;
-                }
-                result = connection.read_datagram() => {
-                    match result {
-                        Ok(data) => {
-                            debug!("Received datagram of size: {}", data.len());
-                            let feedback = parse_feedback(&data);
-                            if let Some(feedback) = feedback {
-                                trace!("Received feedback: {:?}", feedback);
-                                match sender.send(feedback) {
-                                        Ok(_) => {
-                                            debug!("Feedback sent to sender.");
-                                        }
-                                        Err(e) => {
-                                            debug!("Failed to send feedback: {:?}", e);
-                                        }
-                                    }
-                            } else {
-                                debug!("Failed to parse feedback.");
-                            }
-                        }
-                        Err(e) => {
-                            debug!("Error receiving datagram: {e}");
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
     /// Attempts to create a new connection to the specified `peer` address.
     ///
     /// If the connection is successful, the state is updated to `Active`.
@@ -328,7 +222,7 @@ impl ConnectionWorker {
                     Ok(connection) => {
                         let cancel = self.cancel.clone();
                         if let Some(feedback_sender) = &self.feedback_sender {
-                            let recv_task = tokio::spawn(Self::recv_datagrams(
+                            let recv_task = tokio::spawn(recv_tpu_feedback(
                                 connection.clone(),
                                 cancel,
                                 feedback_sender.clone(),
